// src/types/timesheet/model/timesheet-entry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zero consult timesheet - OpenAPI 3.2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TimesheetStatus } from './timesheet-status';

export interface TimesheetEntry {
    'date': string;
    'startTime': string;
    'endTime': string;
    'employeeId': string;
    'customerId': string;
    'id'?: string;
    'createdAt'?: number;
    'description'?: string;
    'status'?: TimesheetStatus;
}